backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
python-dotenv
requests
psycopg2-binary